moving transaction state to sep hook
