+++ src/nft/components/bag/BagFooter.tsx
@@ -20,7 +20,6 @@ import { AllowanceState } from 'hooks/usePermit2Allowance'
 import { useStablecoinValue } from 'hooks/useStablecoinPrice'
 import { useTokenBalance } from 'lib/hooks/useCurrencyBalance'
 import tryParseCurrencyAmount from 'lib/utils/tryParseCurrencyAmount'
-import { useSendTransaction, useTransactionResponse } from 'nft/hooks'
 import { useBag } from 'nft/hooks/useBag'
 import { useBagTotalEthPrice } from 'nft/hooks/useBagTotalEthPrice'
 import useDerivedPayWithAnyTokenSwapInfo from 'nft/hooks/useDerivedPayWithAnyTokenSwapInfo'
@@ -28,11 +27,12 @@ import { useFetchAssets } from 'nft/hooks/useFetchAssets'
 import usePayWithAnyTokenSwap from 'nft/hooks/usePayWithAnyTokenSwap'
 import usePermit2Approval from 'nft/hooks/usePermit2Approval'
 import { PriceImpact, usePriceImpact } from 'nft/hooks/usePriceImpact'
+import { useSubscribeTransactionState } from 'nft/hooks/useSubscribeTransactionState'
 import { useTokenInput } from 'nft/hooks/useTokenInput'
 import { useWalletBalance } from 'nft/hooks/useWalletBalance'
-import { BagItemStatus, BagStatus, RouteResponse, TxStateType } from 'nft/types'
+import { BagStatus } from 'nft/types'
 import { ethNumberStandardFormatter, formatWeiToDecimal, recalculateBagUsingPooledAssets } from 'nft/utils'
-import { PropsWithChildren, useEffect, useMemo, useRef, useState } from 'react'
+import { PropsWithChildren, useEffect, useMemo, useState } from 'react'
 import { AlertTriangle, ChevronDown } from 'react-feather'
 import { useToggleWalletModal } from 'state/application/hooks'
 import { InterfaceTrade, TradeState } from 'state/routing/types'
@@ -290,7 +290,7 @@ const PENDING_BAG_STATUSES = [
 export const BagFooter = ({ setModalIsOpen, eventProperties }: BagFooterProps) => {
   const toggleWalletModal = useToggleWalletModal()
   const theme = useTheme()
-  const { account, chainId, connector, provider } = useWeb3React()
+  const { account, chainId, connector } = useWeb3React()
   const connected = Boolean(account && chainId)
   const totalEthPrice = useBagTotalEthPrice()
   const shouldUsePayWithAnyToken = usePayWithAnyTokenEnabled()
@@ -307,17 +307,13 @@ export const BagFooter = ({ setModalIsOpen, eventProperties }: BagFooterProps) =
     bagStatus,
     setBagExpanded,
     setBagStatus,
-    setLocked: setBagLocked,
-    reset: resetBag,
   } = useBag(
-    ({ itemsInBag, isLocked, bagStatus, setBagExpanded, setBagStatus, setLocked, reset }) => ({
+    ({ itemsInBag, isLocked, bagStatus, setBagExpanded, setBagStatus }) => ({
       itemsInBag,
       isLocked,
       bagStatus,
       setBagExpanded,
       setBagStatus,
-      setLocked,
-      reset,
     }),
     shallow
   )
@@ -327,49 +323,9 @@ export const BagFooter = ({ setModalIsOpen, eventProperties }: BagFooterProps) =
   const usingPayWithAnyToken = !!inputCurrency && shouldUsePayWithAnyToken && chainId === SupportedChainId.MAINNET
 
   const itemsInBag = useMemo(() => recalculateBagUsingPooledAssets(uncheckedItemsInBag), [uncheckedItemsInBag])
-  const sendTransaction = useSendTransaction((state) => state.sendTransaction)
-  const transactionState = useSendTransaction((state) => state.state)
-  const setTransactionState = useSendTransaction((state) => state.setState)
-  const transactionStateRef = useRef(transactionState)
-  const [setTransactionResponse] = useTransactionResponse((state) => [state.setTransactionResponse])
 
-  const purchaseAssets = async (routingData: RouteResponse, purchasingWithErc20: boolean) => {
-    if (!provider || !routingData) return
-    const purchaseResponse = await sendTransaction(
-      provider?.getSigner(),
-      itemsInBag.filter((item) => item.status !== BagItemStatus.UNAVAILABLE).map((item) => item.asset),
-      routingData,
-      purchasingWithErc20
-    )
-    if (
-      purchaseResponse &&
-      (transactionStateRef.current === TxStateType.Success || transactionStateRef.current === TxStateType.Failed)
-    ) {
-      setBagLocked(false)
-      setModalIsOpen(false)
-      setTransactionResponse(purchaseResponse)
-      setBagExpanded({ bagExpanded: false })
-      resetBag()
-    }
-  }
-
-  useEffect(() => {
-    useSendTransaction.subscribe((state) => (transactionStateRef.current = state.state))
-  }, [])
-
-  useEffect(() => {
-    if (transactionStateRef.current === TxStateType.Confirming) setBagStatus(BagStatus.PROCESSING_TRANSACTION)
-    if (transactionStateRef.current === TxStateType.Denied || transactionStateRef.current === TxStateType.Invalid) {
-      if (transactionStateRef.current === TxStateType.Invalid) setBagStatus(BagStatus.WARNING)
-      else setBagStatus(BagStatus.CONFIRM_REVIEW)
-      setTransactionState(TxStateType.New)
-
-      setBagLocked(false)
-      setModalIsOpen(false)
-    }
-  }, [setBagLocked, setBagStatus, setModalIsOpen, setTransactionState])
-
-  const fetchAssets = useFetchAssets(itemsInBag, purchaseAssets)
+  useSubscribeTransactionState(setModalIsOpen)
+  const fetchAssets = useFetchAssets(itemsInBag)
 
   const parsedOutputAmount = useMemo(() => {
     return tryParseCurrencyAmount(formatEther(totalEthPrice.toString()), defaultCurrency ?? undefined)
